mvp/exp/requirements.txt
backoff==2.2.1
pandas==2.2.3
prefect==3.1.14
protobuf==5.29.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
google-generativeai==0.8.4
